refactor(auth): move auth API into feature
