+++ frontend/src/features/tasks/components/TaskList.tsx
@@ -3,7 +3,7 @@ import "./TaskList.css"
 import { Pencil, Trash2, LogOut } from "lucide-react";
 import { useEffect, useState, useCallback } from "react";
 import type { Task } from "../types/task";
-import { getMe } from "../../../lib/api";
+import { getMe } from "../../auth/api/authApi";
 import { fetchTasks, toggleTaskComplete, deleteTask } from "../api/tasksApi";
 import { TaskAdd } from "./TaskAdd";
 import EditTaskModal from "./EditTaskModal"
